Extract generators for solr 4 and 5, and default to installing solr 5
